修复FFmpeg播放的一些问题，并手动集成SMTC (#1386)
* 修复播放进度条的问题

* 细节调整，添加刷新按钮

* 处理SMTC

+++ src/Adapter/Adapter.Implementation/UserAdapter.cs
@@ -151,7 +151,7 @@ namespace Bili.Adapter
         }
 
         /// <inheritdoc/>
-        public UserProfile ConvertToUserProfile(int userId, string userName, string avatar, AvatarSize avatarSize)
+        public UserProfile ConvertToUserProfile(long userId, string userName, string avatar, AvatarSize avatarSize)
         {
             var size = int.Parse(avatarSize.ToString().Replace("Size", string.Empty));
             var image = _imageAdapter.ConvertToImage(avatar, size, size);
